feat: add monolith docker image with realtime server
