+++ .env.example
@@ -37,6 +37,16 @@ NEXT_PUBLIC_REALTIME_PATH=/socket.io
 SOCKET_PATH=/socket.io
 EVENT_PATH=/events
 PORT=4001
+REALTIME_BASE_PATH=/realtime
+# Internal origin is used when the Next.js server emits events to the realtime
+# process. When both run inside one container the default 127.0.0.1 origin is
+# suitable. Override this if the combined image should talk to an external
+# realtime instance.
+REALTIME_INTERNAL_ORIGIN=http://127.0.0.1:3000
+# Optionally override the publicly exposed origin that the browser should use
+# when connecting to the realtime endpoint. Leave unset to fall back to the
+# relative REALTIME_BASE_PATH.
+#REALTIME_PUBLIC_ORIGIN=https://example.com
 
 # ------------------------------------------------------------------------------
 # Deployment targets
