+++ Dockerfile.monolith
@@ -1,0 +1,51 @@
+# Production Dockerfile that bundles the Next.js app and realtime server into a single image
+FROM node:20-slim AS deps
+ENV PNPM_HOME=/root/.local/share/pnpm \
+    PATH=/root/.local/share/pnpm:$PATH \
+    NEXT_TELEMETRY_DISABLED=1
+RUN apt-get update -y && apt-get install -y openssl && rm -rf /var/lib/apt/lists/* \
+    && corepack enable
+WORKDIR /app
+COPY package.json pnpm-lock.yaml ./
+RUN SKIP_PRISMA_POSTINSTALL=1 PRISMA_SKIP_POSTINSTALL_GENERATE=1 pnpm install --frozen-lockfile
+
+FROM node:20-slim AS builder
+ENV PNPM_HOME=/root/.local/share/pnpm \
+    PATH=/root/.local/share/pnpm:$PATH \
+    NODE_ENV=production \
+    NEXT_TELEMETRY_DISABLED=1
+RUN apt-get update -y && apt-get install -y openssl && rm -rf /var/lib/apt/lists/* \
+    && corepack enable
+WORKDIR /app
+COPY --from=deps /app/node_modules ./node_modules
+COPY . .
+RUN pnpm prisma:generate
+RUN pnpm build
+
+FROM node:20-slim AS runner
+ENV PNPM_HOME=/root/.local/share/pnpm \
+    PATH=/root/.local/share/pnpm:$PATH \
+    NODE_ENV=production \
+    NEXT_TELEMETRY_DISABLED=1 \
+    PORT=3000 \
+    REALTIME_BASE_PATH=/realtime \
+    NEXT_PUBLIC_REALTIME_URL=/realtime \
+    NEXT_PUBLIC_REALTIME_PATH=/realtime/socket.io \
+    REALTIME_SERVER_EVENT_PATH=/realtime/events \
+    REALTIME_INTERNAL_ORIGIN=http://127.0.0.1:3000 \
+    REALTIME_SERVER_URL=http://127.0.0.1:3000/realtime
+RUN apt-get update -y && apt-get install -y openssl && rm -rf /var/lib/apt/lists/* \
+    && corepack enable
+WORKDIR /app
+COPY --from=builder /app/node_modules ./node_modules
+COPY --from=builder /app/package.json ./package.json
+COPY --from=builder /app/pnpm-lock.yaml ./pnpm-lock.yaml
+COPY --from=builder /app/.next ./.next
+COPY --from=builder /app/public ./public
+COPY --from=builder /app/prisma ./prisma
+COPY --from=builder /app/scripts ./scripts
+COPY --from=builder /app/realtime-server ./realtime-server
+
+EXPOSE 3000
+
+CMD ["sh", "-c", "pnpm prisma migrate deploy && node scripts/start-combined-server.mjs"]
